% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agg60s.R
\name{agg60s}
\alias{agg60s}
\title{agg60s}
\usage{
agg60s(data = c(), epoch = c())
}
\arguments{
\item{data}{Data frame including time stamps and step metrics to be aggregated}

\item{epoch}{Epoch length in which the original data are sampled}
}
\value{
New data frame aggregated per 60s
}
\description{
agg60s
}
