% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_midnights.R
\name{detect_midnights}
\alias{detect_midnights}
\title{detect_midnights}
\usage{
detect_midnights(t = c())
}
\arguments{
\item{t}{Vector of timestamps in which look for midnights}
}
\value{

}
\description{
detect_midnights
}
