% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_daily.R
\name{get_daily}
\alias{get_daily}
\title{get_daily}
\usage{
get_daily(datadir_scriin, datadir_fibion, outputdir)
}
\arguments{
\item{datadir_scriin}{Directory where the scriin files are stored}

\item{datadir_fibion}{Directory where the fibion files are stored}

\item{outputdir}{Directory to store the outpu}
}
\value{

}
\description{
get_daily
}
