% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_timeseries.R
\name{get_timeseries}
\alias{get_timeseries}
\title{get_timeseries}
\usage{
get_timeseries(id, day, scriin_day, fibion_day, outputdir, from = 9, to = 18)
}
\arguments{
\item{id}{id of participant}

\item{day}{day being analyzed}

\item{scriin_day}{data from scriin}

\item{fibion_day}{data from fibion}

\item{outputdir}{Directory with output files}

\item{from}{Numeric, from what time (hours, accept decimals)}

\item{to}{Numeric, to what time (hours, accept decimals)}
}
\value{

}
\description{
get_timeseries
}
