% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_daily.R
\name{merge_daily}
\alias{merge_daily}
\title{merge_daily}
\usage{
merge_daily(datadir_scriin, datadir_fibion, outputdir, save_timeseries = TRUE)
}
\arguments{
\item{datadir_scriin}{Directory where the scriin files are stored}

\item{datadir_fibion}{Directory where the fibion files are stored}

\item{outputdir}{Directory to store the outpu}
}
\value{

}
\description{
merge_daily
}
