% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFile.R
\name{readFile}
\alias{readFile}
\title{readFile}
\usage{
readFile(file = c(), mon = c())
}
\arguments{
\item{file}{Character with the full path to the file to be read}

\item{mon}{Character indicating the monitor used, options are FIBION, Scriin, ActiGraph, Other}
}
\value{
Data frame with the accelerometer data information
}
\description{
readFile
}
