% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.R
\name{resample}
\alias{resample}
\title{resample}
\usage{
resample(
  data = c(),
  timestamp_col = c(),
  timestamp_format = c(),
  id_col = c(),
  mon = c()
)
}
\arguments{
\item{data}{Data frame including time stamps and step metrics to be resampled}

\item{timestamp_col}{Character or numeric indicating the column where the timestamp is stored}

\item{timestamp_format}{Character string giving a date-time format as used by \link[base]{strptime}.}

\item{id_col}{Character or numeric indicating the column where the ID is stored. Leave blank if the ID is not stored}

\item{mon}{Character indicating what monitor was used. Options are Fibion, Scriin, ActiGraph, Other.}
}
\value{
Data frame resampled
}
\description{
resample
}
