% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timestamp2iso8601.R
\name{timestamp2iso8601}
\alias{timestamp2iso8601}
\title{timestamp2iso8601}
\usage{
timestamp2iso8601(x, tz = "", time_format = c())
}
\arguments{
\item{x}{Character vector with timestamps}

\item{tz}{Character, timezone in which the data were collected as in http://en.wikipedia.org/wiki/Zone.tab}

\item{time_format}{Character string giving a date-time format as used by \link[base]{strptime}.}
}
\value{

}
\description{
timestamp2iso8601
}
